app/requirements.txt
bottle==0.12.19
psycopg2-binary==2.9.5
redis==4.5.1
requests==2.28.2
resend==0.6.0 